requirements.txt
fastapi
fastapi-mcp
uvicorn
pandas
mcp_proxy
python-dotenv

langchain_openai
langchain
langchain_community
langgraph
openai
pgvector
psycopg2-binary

openai
anthropic

pydantic-settings